+++ CHANGELOG.rst
@@ -5,6 +5,9 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+`0.8`_ (2019-02-07)
+~~~~~~~~~~~~~~~~~~~
+
 - Removed ``apps_urlconf_for_site`` and ``apps_middleware`` since
   feincms3's ``apps_middleware`` now automatically does the right thing
   when used after ``site_middleware``.
@@ -101,4 +104,5 @@ Change log
 .. _0.5: https://github.com/matthiask/feincms3-sites/compare/0.4...0.5
 .. _0.6: https://github.com/matthiask/feincms3-sites/compare/0.5...0.6
 .. _0.7: https://github.com/matthiask/feincms3-sites/compare/0.6...0.7
-.. _Next version: https://github.com/matthiask/feincms3-sites/compare/0.7...master
+.. _0.8: https://github.com/matthiask/feincms3-sites/compare/0.7...0.8
+.. _Next version: https://github.com/matthiask/feincms3-sites/compare/0.8...master